Finish refactor of league_stats_module_test with helper method tests; write any missing since refactor to use maxmin
